// src/generated/models/PowerPlatformforAdminsV2Model.ts
export interface AdvisorRecommendation {
  // The recommendation name.
  scenario?: string;
  details?: AdvisorRecommendationDetails;
}

export interface AdvisorRecommendationDetails {
  // The number of resources
  resourceCount?: number;
  // Time when the recommendation was refreshed
  lastRefreshedTimestamp?: string;
  // Time when the recommendation will be refreshed again
  expectedNextRefreshTimestamp?: string;
}
// ...
export interface BillingInstrumentModel {
  // The tenant subscription Id.
  subscriptionId?: string;
  // The resource group within the tenant subscription.
  resourceGroup?: string;
  id?: string;
}
// ...
export interface BillingPolicyResponseModel {
  id?: string;
  name?: string;
  status?: BillingPolicyStatus;
  location?: string;
  billingInstrument?: BillingInstrumentModel;
  createdOn?: string;
  createdBy?: Principal;
  lastModifiedOn?: string;
  lastModifiedBy?: Principal;
}

export interface BillingPolicyResponseModelResponseWithOdataContinuation {
  value?: BillingPolicyResponseModel[];
  "@odata.nextLink"?: string;
}

export type BillingPolicyStatus = string;
// ...
export interface Connection {
  name?: string;
  id?: string;
  type?: string;
  properties?: {
    apiId?: string;
    displayName?: string;
    iconUri?: string;
    statuses?: ConnectionStatus[];
    connectionParametersSet?: {
      name?: string;
      values?: Record<string, unknown>;
    };
    connectionParameters?: Record<string, unknown>;
    keywordsRemaining?: number;
    isSsoConnection?: boolean;
    createdBy?: ConnectionCreatedBy;
    createdTime?: string;
    lastModifiedTime?: string;
    expirationTime?: string;
    testLinks?: {
     requestUri?: string;
     method?: string;
   }[];
    environment?: {
      id?: string;
      name?: string;
    };
    accountName?: string;
    metadata?: Record<string, unknown>;
    allowSharing?: boolean;
  };
}

export interface ConnectionCreatedBy {
  id?: string;
  displayName?: string;
  email?: string;
  type?: string;
  tenantId?: string;
  userPrincipalName?: string;
}
// ...
export interface ConnectionStatus {
  status?: string;
  target?: string;
  error?: {
    code?: string;
    message?: string;
  };
}
// ...
export type CrossTenantConnectionconnectionType = 'Inbound'|'Outbound';

export interface CrossTenantConnection {
  // The Azure AD tenant ID to or from which the cross-tenant connection occurred.
  tenantId: string;
  // The direction of the cross-tenant connection.
  connectionType: CrossTenantConnectionconnectionType;
}

export type CrossTenantConnectionReportstatus = 'Received'|'InProgress'|'Completed'|'Failed';

export interface CrossTenantConnectionReport {
  // The Azure AD tenant ID for which the report was generated.
  tenantId: string;
  // The report ID.
  reportId: string;
  // The date when the cross-tenant connection report was requested.
  requestDate: string;
  // The start of the report date window.
  startDate: string;
  // The end of the report date window.
  endDate: string;
  status: CrossTenantConnectionReportstatus;
  // The page of cross-tenant connections occurring within the report date window.
  connections?: CrossTenantConnection[];
  // Next page URI for additional cross-tenant connections.
  "@odata.nextLink"?: string;
}
// ...
export interface EnvironmentGroup {
  id?: string;
  displayName?: string;
  description?: string;
  parentGroupId?: string;
  childrenGroupIds?: string[];
  createdTime?: string;
  createdBy?: Principal;
  lastModifiedTime?: string;
  lastModifiedBy?: Principal;
}
// ...
export interface GetConnectorByIdResponse {
  name?: string;
  id?: string;
  type?: string;
  properties?: {
    displayName?: string;
    iconUri?: string;
    iconBrandColor?: string;
    apiEnvironment?: string;
    isCustomApi?: boolean;
    blobUrisAreProxied?: boolean;
    runtimeUrls?: string[];
    primaryRuntimeUrl?: string;
    doNotUseApiHubNetRuntimeUrl?: string;
    metadata?: {
      source?: string;
      brandColor?: string;
      allowSharing?: boolean;
      useNewApimVersion?: string;
      version?: {
        previous?: string;
        current?: string;
      };
    };
    capabilities?: string[];
    interfaces?: Record<string, unknown>;
    description?: string;
    createdTime?: string;
    changedTime?: string;
    releaseTag?: string;
    tier?: string;
    publisher?: string;
    rateLimit?: number;
    apiVersion?: string;
  };
}
// ...
export interface Principal {
  id?: string;
  displayName?: string;
  email?: string;
  type?: string;
  tenantId?: string;
  userPrincipalName?: string;
}
// ...
export type WebsiteDtotemplateName = 'DefaultPortalTemplate'|'PowerPortals_ProgramRegistration'|'PowerPortals_BookMeeting';
export type WebsiteDtopackageInstallStatus = 'None'|'Installed'|'Uninstalled'|'InstallRequested'|'UninstallRequested'|'InstallFailed'|'UninstallFailed'|'Installing'|'Uninstalling'|'InstallScheduled'|'InstallRetrying'|'TemplateInstalled';
export type WebsiteDtotype = 'Trial'|'Production';
export type WebsiteDtostatus = 'OperationComplete'|'OperationInProgress'|'OperationNotStarted'|'OperationFailed';
export type WebsiteDtositeVisibility = 'public'|'private';

export interface WebsiteDto {
  // Website unique identifier (ID)
  id?: string;
  // Website name
  name?: string;
  // Website creation time in the ISO 8601 UTC format
  createdOn?: string;
  // Website template name
  templateName?: WebsiteDtotemplateName;
  // Website URL
  websiteUrl?: string;
  // Tenant unique identifier (ID) of the website
  tenantId?: string;
  // Organization URL of the website
  dataverseInstanceUrl?: string;
  // Environment name of the website
  environmentName?: string;
  // Environment unique identifier (ID) of the website
  environmentId?: string;
  // Organization unique identifier (ID) of the website
  dataverseOrganizationId?: string;
  // Language ID - https://go.microsoft.com/fwlink/?linkid=2208135
  selectedBaseLanguage?: number;
  // Custom hostnames added for the website
  customHostNames?: string[];
  // Dataverse record unique identifier (ID) of the website
  websiteRecordId?: string;
  // Subdomain of website
  subdomain?: string;
  // Package installation status of the website
  packageInstallStatus?: WebsiteDtopackageInstallStatus;
  // Application type of the website
  type?: WebsiteDtotype;
  // Time (in days) to expiration of the website
  trialExpiringInDays?: number;
  // Time (in days) to website deletion, if suspended
  suspendedWebsiteDeletingInDays?: number;
  // Package version of the website
  packageVersion?: string;
  // Website eligibility for early upgrade
  isEarlyUpgradeEnabled?: boolean;
  // Custom error enablement for Website
  isCustomErrorEnabled?: boolean;
  // Entra ID (formerly Azure Active Directory) object unique identifier (ID)
  applicationUserAadAppId?: string;
  // User unique identifier (ID) of the website owner
  ownerId?: string;
  // Website status
  status?: WebsiteDtostatus;
  // Website visibility status
  siteVisibility?: WebsiteDtositeVisibility;
}